ENH: SlicerQT - Added method 'extractModuleName' to qSlicerLoadableModuleFactory. Also added test qSlicerLoadableModuleFactoryTest1
Currently, modules are uniquely identified using their library name.
This later one is platform specific. For example, ThresholdLib.dll on windows
or libThresholdLib.so on linux.

This method allows to return the same identified in all cases.
In this example, 'threshold' will be returned.

The method extractModuleName is just added and tested, it's not yet used by
the module manager.

git-svn-id: http://svn.slicer.org/Slicer4/branches/Slicer4@12731 3bd1e089-480b-0410-8dfb-8563597acbee