ENH: Add list of "additional module paths" to "module settings panel"
The qSlicerDirectoryListView has been integrated to the panel and
provides the user with the ability to remove/add paths that will be
used to configure the different module factories.

Each factory will then scan this list of additional path and attempt
to load extra modules if any.

Each time a path is removed/added to the list, a red label entitled
"* Restart requested" is shown to the user.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@16946 3bd1e089-480b-0410-8dfb-8563597acbee