BUG: fix color by orientation mode in tensor display
Issue was that output scalar type, which could be either float or
unsigned char depending on the operation being used, was being
used to cast the pointer to the input tensors.  Apparently this
output type information was previously not used to define the
type of the input data pointer (prior to vtk 5.8) but with the 
current version the output pixel type as defined in the 
request information method is being respected.


git-svn-id: http://svn.slicer.org/Slicer4/trunk@17748 3bd1e089-480b-0410-8dfb-8563597acbee