BUG: patch to remove messages upon exit about unknown class vtkInternalLightBoxManagerProxy
vtkStandardNewMacro does not like it when the specified class is internal
to another class. This is because the preprocessor needs to use the
classname as a return type and a classname where the former needs to be
fully scoped and the latter only needs the internal class.

We provide the equivalent to what the StandardNewMacro should have done.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@18532 3bd1e089-480b-0410-8dfb-8563597acbee