BUG: fixing bug 1764
The problem was caused by the following. When the scene closes, parameter node
used by CropVolume is removed. The new node has never been created, since it
is done in setMRMLScene(). To resolve this, I added observer for SceneEndClose
event, where I initialize a new parameter node. In the process of fixing this,
the API of the parameter node has been simplified so that only IDs, not
pointers to the actual nodes are kept. ChangeTracker module needs to be
updated to account for this change.


git-svn-id: http://svn.slicer.org/Slicer4/trunk@19473 3bd1e089-480b-0410-8dfb-8563597acbee