BUG: Don't assert when loading/restoring display node with volume already in scene
When loading or restoring a scene, GetDisplayScalarRange() can be called
while updating the display node. A sanity check makes sure that if the
display node doesn't have an image data, its volume node neither. This is
a true statement except when loading a scene, the volume node might exist
already in the existing scene and contain an imagedata but the display node
being loaded (ReadXMLAttributes) has not yet the image data.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@19600 3bd1e089-480b-0410-8dfb-8563597acbee