ENH: Ensure extensionsInstallPath is properly set on MacOSX
* Since extension libraries and executable are fixed up, there are expected
to be located within the "Slicer.app/Contents" folder. For that purpose,
the following changes have been introduced:

 -  Introduce method "defaultExtensionsInstallPathForMacOSX" in the pimpl
of qSlicerCoreApplication. This method is used to force the extensionsInstallPath
to have always the same value on macosx.

 - The ExtensionsInstallPathButton in extension settings panel is now
always disabled if the operating system is macosx.

 - In case of non-installed Slicer on macosx, the path prefixed with
"bin/Slicer.app/Contents/<Slicer_BUNDLE_EXTENSIONS_DIRNAME>" are added as search paths.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@19648 3bd1e089-480b-0410-8dfb-8563597acbee