ENH: Add loading options to MRML Scene file
Option to clear the scene (equivalent to Load Scene) or not (equivalent to Import).
If the scene is imported, then the option Copy Cameras (enabled by default) copy the properties of the scene-to-import camera nodes into the existing nodes with the same name.
Move qSlicerSceneIO into the Data module to add an OptionsWidget to control the options.

Closes #1952

git-svn-id: http://svn.slicer.org/Slicer4/trunk@20135 3bd1e089-480b-0410-8dfb-8563597acbee