BUG: Un-install extensions after Slicer restarts.
Extensions are now scheduled to be un-installed. When doing so the extension name
will be added to list in the settings identified by "Extensions/ScheduledForUninstall"
and the extensions will be effectively uninstalled when Slicer restarts.

User has also the possibility to cancel the extension scheduled for uninstall.

Note that the uninstall mechanism does NOT yet account for extension dependencies.

See http://www.na-mic.org/Bug/view.php?id=1906

Closes #1906

git-svn-id: http://svn.slicer.org/Slicer4/trunk@20156 3bd1e089-480b-0410-8dfb-8563597acbee