ENH: Default layout is 4-Up
The main use case for Slicer is to load volumes, the 2D slice views being
the best views to visualize the volumes, the 3D view being useless at first,
the best default layout should give as much space for the 2D slice views,
the 4-up layout is ideal as a first default layout.
Note, the layout is saved before exiting and restored next time the
application is started so the initial layout is not critical.

git-svn-id: http://svn.slicer.org/Slicer4/branches/Slicer-4-1-2@20425 3bd1e089-480b-0410-8dfb-8563597acbee