ENH: Use checkbox to control the "Expanded" hierarchy node property
The "Expanded" property is a property of the model and not of the view.
This is why the "Collapsed/Expanded" tree view visual feedback can't be
used to control the model, it was linked to the view, not the underlying
model.

Note that the Visibility of models is still not correctly working.

Issue #2163

git-svn-id: http://svn.slicer.org/Slicer4/trunk@20768 3bd1e089-480b-0410-8dfb-8563597acbee