BUG: Write vtkMRMLVolumePropertyNode functions with highest precision
The point precision can be very important. Especially when have a
threshold like transfer function:
      _____
     |     |
_____|     |____
0   1,2   3,4   5
in that case, the points 1 and 2 have almost the same 'x'. But it is a
different 'x' otherwise the points would be merged.
A difference of std::numeric_limits<double>::epsilon() can be used.
Issue #2485

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21048 3bd1e089-480b-0410-8dfb-8563597acbee