STYLE: Remove vtkMRMLXXXColorYYYNode::GetTypeAsIDString()
a) there was a duplication of information between "IDs" and "GetTypeAsString"
b) A color node should not know about how IDs are set.
It doesn't even set the SingletonTag to itself,
however vtkMRMLColorLogic is the one that sets the ID
(by setting the singleton), so vtkMRMLColorLogic should be the one
that knows about IDs.
Actually, to be really consistent, only the scene knows about IDs
(see vtkMRMLScene::GenerateUniqueID).

Closes #2613

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21111 3bd1e089-480b-0410-8dfb-8563597acbee