BUG: fix scene view crashes in 2514
The layout node in the layout logic was not cleared when a scene
was restored, which meant that it was out of date with respect
to the current scene and led to a crash.  Now we explicitly unreference
it when a scene restore begins and then re-access it after
the scene has been restored.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21176 3bd1e089-480b-0410-8dfb-8563597acbee