BUG: issue #2513 avoid a crash when restoring scene view
This change probably speeds things up to - the issue is that without
this change items in the model are being updated while the scene
is in an inconsistent state.  From testing by me and Nicole we saw
no side effects of this call and it allows the brain atlas to
run to completion, although it still crashes when closing the scene
(in a SceneView destructor) which means I believe it is now
the same as issue #2512.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21354 3bd1e089-480b-0410-8dfb-8563597acbee