COMP: Use VTK 5.10.1 to include latest bug fixes for 5.10.1
A new branch named "slicer-v5.10.1" has been created on the Slicer VTK fork.
It is based of VTK v5.10.1 and includes Slicer specific patches.

List of commit specific to v5.10.1 is listed here:
  https://github.com/Kitware/VTK/compare/v5.10.0...v5.10.1

List of commit specific to Slicer is listed here:
  https://github.com/Slicer/VTK/compare/Kitware:v5.10.1...Slicer:slicer-v5.10.1

Thanks to David Gobbi and Csaba Pinter for suggesting the update.

Fixes #2950

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21703 3bd1e089-480b-0410-8dfb-8563597acbee