ENH: Added versioning of SetupForDevelopment script
Recording the version of the development environment will help notify the
developer to re-run the SetupForDevelopment script. This can be done when there
are changes made to the environment setup scripts or to the client-side hooks.

The SetupForDevelopment script records the version of the developer setup in the
Git config file. This version is tested against the version specified in the
SetupForDevelopment script using the pre-commit hook.

This means, if there is a change in the developer setup (hooks, etc.), the
version number in the SetupForDevelopment script should be incremented to
notify other developers that they should update their setup.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21913 3bd1e089-480b-0410-8dfb-8563597acbee