ENH: expose module instance and module widget python objects
With this change it's possible to directly access the module widget
of a scripted module as slicer.modules.<module>Widget (i.e. the
actual instance that is visible when the module is selected).  This
makes it possible to access internal variables for debugging or
to access widget and logic instances for self-tests or automation.

For symmetry slicer.modules.<module>Instance is also exposed,
although there is generally not anything useful in that class
instance since it is primarily declaritive.  But sometimes it can
have global helper methods, so making it available can be useful.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21973 3bd1e089-480b-0410-8dfb-8563597acbee