BUG: Set the color range when changing number of colors
When copying a color table node and reducing the number
of colors in it, the color table range wasn't updated.
This led to problems in the Editor module when the selected
drawing integer wasn't matched up to the expected color.
Now when the number of colors is set on a user editable
color node, the range is set on the color look up table
and on the GUI slider at the same time.
Also updated the number of colors spin box to only trigger
a reset of the number of colors on the node on the editingFinished signal
as othewise cutting down a table from 300 to 50 colors could
empty out the table and refill it with black rather than 
creating a subset of the original table.

Issue #3115



git-svn-id: http://svn.slicer.org/Slicer4/trunk@21993 3bd1e089-480b-0410-8dfb-8563597acbee