ENH: add an attribute indicating the layout name of the view of anno manip
When a widget is actively moving a fiducial, the viewer should not
be modified by external commands, such as chaning the slice node
since that will reset the widget.  To avoid this, the added attribute
allows other code to know which view is the one in which the widget
is active, and can thereby keep it out of any other updates.

(See Landmark registration project for an example
gitub.com/pieper/LandmarkRegistration)

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22011 3bd1e089-480b-0410-8dfb-8563597acbee