ENH: Only one unit per quantity accessible to the user
Now only one unit per quantity is accessible to the user. This unit can
be customized and it's logged in the settings so it can be restored
when opening the application.
On the implementation side of things:
- Add access method to all the referenced unit nodes in the selection node
- Remove the unit widget
- Hide module
- Create the unit widget to easily customize units (add the plugins with
it)
- Create the unit settings widget using the unit widget. This widget
is tailored to be used in the setting panel (no plugin)
- Refactorize the setting panel to use the setting unit widget and
log the unit node properties
- Create only node for each quantity exposed to the user. The preset nodes
are kept in a special scene in the logic.

See Issue #1694

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22056 3bd1e089-480b-0410-8dfb-8563597acbee