BUG: #2990 use conventional layout if saved layout not available
When a module or extension creates a custom layout the number (say, 100)
is saved in the settings file and slicer tries to restore that
on the next start up.  However these layouts can be transient creations
and won't be available at startup.  This patch improves handling of
that case - rather than printing a warning and showing a blank screen, the
program now starts with the regular layout which leads to less
user confusion.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22142 3bd1e089-480b-0410-8dfb-8563597acbee