BUG: fix crash on exit for py_MarkupsSceneViewRestoreTestManyLists
On some runs of the py_MarkupsSceneViewRestoreTestManyLists test, the
volume rendering displayable manager would trigger a segmentation
fault when it was being deleted on exit. It came from requesting
a render after setting and observing the NULL displayable node.
Added a test for a null node and if true don't request a render.
Tested via the GUI and ran the volume rendering tests.

From: naucoin <naucoin@3bd1e089-480b-0410-8dfb-8563597acbee>

git-svn-id: http://svn.slicer.org/Slicer4/branches/Slicer-4-3@22491 3bd1e089-480b-0410-8dfb-8563597acbee