BUG: #3412 fix crash when no tensors on fiber label select
When the input fiber bundles have no tensors, we don't want to create
an output array.  When and empty tensor array was passed to the vtk
xml reader it crashed.

Now, don't create an output tensor array if there is not one on the input.

From: pieper <pieper@3bd1e089-480b-0410-8dfb-8563597acbee>

git-svn-id: http://svn.slicer.org/Slicer4/branches/Slicer-4-3@22500 3bd1e089-480b-0410-8dfb-8563597acbee