BUG: fix updating fiducials after scene load
After loading a scene, changing the display properties of
a fiducial list was not being reflected in the 2d and 3d
views. There was a missing event observation in the
Markups displayable managers, adding in 
vtkMRMLDisplayableNode::DisplayModifiedEvent
to the list of events observed on a markups node
fixed the problem.

Issue #3480



git-svn-id: http://svn.slicer.org/Slicer4/trunk@22800 3bd1e089-480b-0410-8dfb-8563597acbee