ENH: Improve readability using vtkNew where possible
In addition to be shorter, it simply conveys the idea of locally
instantiating an object.

Indeed, vtkNew allow to instantiate an object on the heap and deleting it
when it goes out of scope if its reference count has not been incremented by
either assigning the object to a smart pointer or calling Register.

This commit continues the work initiated in commit r22539

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22896 3bd1e089-480b-0410-8dfb-8563597acbee