ENH: Create new templates for wizard
Refactor the ModuleWizard.py templates for use by the new
SlicerWizard.py. The main changes are that the module templates no
longer have extension templates wrapped around them (and the extension
templates no longer contain modules), the directory organization has
been improved, and the placeholders are all now TemplateKey.

This last change in particular was to allow the SuperBuild template to
be named "SuperBuild" (which is easy for users to type and logical given
the new organization), and still contain files named like "SuperBuild"
that should not be renamed.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22947 3bd1e089-480b-0410-8dfb-8563597acbee