ENH: Confirm changes to extension information
Change how 'extensionwizard.py --publish' modifies the extension
information to prompt the user before making changes to URL's (the
HOMEPAGE in particular, but also the ICONURL) if the existing URL does
not have 'example.com' as the host. As example.com is a reserved name,
such a URL (e.g. the default values from the templates) cannot possible
be valid and so should be safe to change in all cases, but this should
cause us to stop and confirm if we're about to change a URL that the
user has already changed to some potentially valid location (e.g. the
wiki).

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22991 3bd1e089-480b-0410-8dfb-8563597acbee