ENH: Add version information to SlicerWizard
Add package version information to SlicerWizard, and add display of the
same via the usual '--version' argument to the Extension Wizard. Note
that this is done via a committed file containing the version
information which is duplicated from the top-level CMakeLists.txt. There
is also a CMakeLists.txt now in Utilities/Scripts/SlicerWizard to update
this file, which should keep the contents in sync. Updating the version
should ideally be done by making the changes in the top-level
CMakeLists.txt, configuring, and *then* committing the changes (which
will include as a result the SlicerWizard __version__.py).

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23000 3bd1e089-480b-0410-8dfb-8563597acbee