BUG: Separator appears in MRML node combobox menu with custom items only
The MRML node combobox can have extra items on the top or bottom of the node list. There are built-in extra items, such as None, Add, etc. When these are enabled, then a separator appears between the nodes and the items. However, when no built-in extra item is enabled, just custom extra items are added, then the separator did not show up.

From: Csaba Pinter <csaba.pinter@queensu.ca>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23046 3bd1e089-480b-0410-8dfb-8563597acbee