COMP: Fix windows packaging error disabling GitPython and PyGithub
The problem is that the directory structure is too deep. Some path
have a length greater than 260 characters and seems to be an issue
for the CMake install command due to the windows maximum path
length limitation [1]

Waiting this issue is addressed, this commit disables the package on
Windows to ensure packages will be uploaded.

[1] http://msdn.microsoft.com/en-us/library/windows/desktop/aa365247%28v=vs.85%29.aspx#maxpath

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23088 3bd1e089-480b-0410-8dfb-8563597acbee