BUG: Don't rely on URL change to update search
Also set the current search text (used to avoid unnecessary updates)
when applying the search filter. This way, the test still works if we
never get the URL changed signal (which seems to happen on some machines
for reasons unknown). In theory, since the URL change is being triggered
by JavaScript and not waiting on a server response, it should happen
immediately anyway, possibly even synchronous in the call to
evaluateJavaScript().

This should fix inability to clear the search text on some machines
(because failure to update the last search text means we always think
we have no search, and therefore don't need to do anything to clear the
search text).

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23131 3bd1e089-480b-0410-8dfb-8563597acbee