BUG: Fix extension manager item sizing
Refactor qSlicerExtensionsDescriptionLabel to manage the text internally
as our own QTextDocument instance, and to do the painting ourselves.
This gives us greater control over the painting with less work, and
greatly simplifies producing a correct size hint, since we can just ask
the document for its size and get a correct answer regardless of
formatting, rather than trying to compute it ourselves from font
metrics. Now that we have a reasonable size hint, also use it to set the
item size hint, which ensures that the items are sufficiently tall e.g.
in case of the extra text line when an extension is incompatible.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23167 3bd1e089-480b-0410-8dfb-8563597acbee