ENH: Add 'More' link in extension manager (1/2)
Use the newly available extension ID to provide a 'More' link in the
extension manager. Add necessary code to implement link signals in
qSlicerExtensionsDescriptionLabel.

This is part 1 of 2; the links are not yet functional as we currently
lack a way to open the page. (We don't want to use the web view for
installing, as that would cause an undesired switch to the 'install'
tab.)

Note that the QSignalMapper for links is removed, as it is not needed;
the link URL provides the necessary information and can be used
directly.

Issue #3608

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23171 3bd1e089-480b-0410-8dfb-8563597acbee