BUG: Fixed DICOM browser persistence
If the DICOM browser persistent option is enabled (usually when the DICOM browser is displayed on a second monitor) then the browser should be kept open when an image is loaded or the user opens another module.
The persistent flag was observed almost everywhere where it was needed, except at the end of the image loading step.

From: Andras Lasso <lasso@queensu.ca>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23233 3bd1e089-480b-0410-8dfb-8563597acbee