ENH: Allow users to acknowledge all DICOM loading errors with a single click
Problem: Each DICOM loading error was displayed in a separate popup, which
 required the user to click as many times as many series failed to load.
In some cases there are a large number of failures (for example,
experimental MRI sequences), which makes loading of a complete study very
frustrating (need to click 30 times, waiting several seconds between each).

Solution: Changed the error reporting to collect all errors and show it to the user at the end of the loading in a single error popup.

From: Andras Lasso <lasso@queensu.ca>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23234 3bd1e089-480b-0410-8dfb-8563597acbee