ENH: Sank addSeriesInSubjectHierarchy function down to abstract DICOMPlugin
The addSeriesInSubjectHierarchy function (which was in DICOMScalarVolumePlugin so far) adds
the loaded node to subject hierarchy to the proper patient (subject) and study according to
the DICOM tags from the first file referenced in the associated loadable. This function is
now accessible from all the DICOM plugins. It needs to be called after the loaded object is
created in the load() function. Now it is called by DICOMDiffusionVolumePlugin too.

From: Csaba Pinter <csaba.pinter@queensu.ca>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23238 3bd1e089-480b-0410-8dfb-8563597acbee