ENH: Improve styling of wizard actions
Replace the QPushButton extension wizard actions with "flat"
QToolButton's, using (for now) a forward-arrow icon. This gives a
cleaner look than the odd stack of unusually wide buttons, that is
loosely based on Windows wizards (which were the only example of a
list-of-possible-actions - as opposed to list-of-steps - type wizard I
could find).

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23268 3bd1e089-480b-0410-8dfb-8563597acbee