BUG: Pass template path to built wizard
Modify the build tree wrapper script for running the CLI Extension
Wizard to explicitly pass the template path for the built-in templates
(in the source tree). Since this script is configured anyway and only
used for the build tree, this is more straight forward than the trick we
use in the Extension Wizard module of parsing the CMakeCache.txt to find
the source directory.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23295 3bd1e089-480b-0410-8dfb-8563597acbee