BUG: Fix qrcc.py operation on Windows
Change qrcc.py to use a temporary file to receive the raw binary rcc
data on Windows, in order to work around what appears to be a bug in rcc
where it is performing line ending (LF -> CRLF) conversion when writing
to stdout, resulting in corrupt data which causes Qt to crash when
trying to load the resources.

See also https://bugreports.qt-project.org/browse/QTBUG-39422, r23297.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23300 3bd1e089-480b-0410-8dfb-8563597acbee