BUG: Don't stop uninstall on first error
Change how we process scheduled extension uninstall operations to
continue trying to uninstall all extensions, rather than stopping the
process after the first error. The old behavior is a result of operand
order to the '&&' operator, which employs short-circuit evaluation to
skip evaluation of the second argument (the call to uninstall the
extension) if the first argument (the overall result 'accumulator') is
false, and was likely unintentional, due to overlooking the short
circuit evaluation.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23319 3bd1e089-480b-0410-8dfb-8563597acbee