COMP: On MacOSX, waiting issue #3741 is solved uses VTK 5.10
As illustrated in the bug report [1], the maximum number of linked libraries
is reached  on MacOSX 10.6 (see below for an example of message) but there
are no problem on MacOSX 10.8 (at least when targeting 10.9).

// ---------
ld: warning: duplicate dylib /path/to ...
[...]
ld: warning: duplicate dylib /path/to ...
two level namespace mach-o files can link with at most 253 dylibs, this link would use 287 dylibs
// ---------

Waiting we come up with a solution or understand better the scope of the
problem, this commit force the VTK version to 5.10.

[1] http://na-mic.org/Mantis/view.php?id=3741

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23344 3bd1e089-480b-0410-8dfb-8563597acbee