BUG: Consider inner build dir when checking for max root directory length.
This commit introduces the convenience CMake module "PreventDirWithTooManyChars"
and update code to account for length of the inner build subdirectory.

It means if the maximum length of the top-level build directory is 40,
assuming the inner build directory is "Slicer-build", the check in
the inner build will expect a path of length 53.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23361 3bd1e089-480b-0410-8dfb-8563597acbee