BUG: fix logic on showing slice intersections
Added an attribute (and a name) for the volume slice
model display nodes and check for it in a new slice
logic static method. If it's not a volume slice model
display node, use the visibility of the regular
model to determine if should show the slice intersections,
otherwise rely on the slice intersections visible flag.

Issue #3718


git-svn-id: http://svn.slicer.org/Slicer4/trunk@23495 3bd1e089-480b-0410-8dfb-8563597acbee