COMP: Update VTK6 to include support for VTK_ENABLE_KITS. See #3741
This commit updates the version of VTK so that the changes introduced
in commit r23509 can have an effect.

An attempt to cherry-pick only the commits introducing the VTK_ENABLE_KITS
feature has been done. It turned out that the kits feature and the
introduction of the support for different backend were too much tangled.

After reviewing the list of changes, most them includes bug fixes, addition
of VTK tests, fixes for memory initialization, optimization.
More specifically, here is a list of the one that are of interest
for Slicer are:

  kitware/VTK@504036 : Added "sticky" axes mode to vtkCubeAxesActor
  kitware/VTK@196bbe : Refactor GetRange to compute all ranges at the same time.
                       [...] improves the total runtime by 1.5x to 3x.
  kitware/VTK@87291c : Improve ComputeScalarRange when we have a single component.
                       In testing this takes the ComputeScalarRange on 600^3 from ~430 ms to ~20ms.
  kitware/VTK@cbd488 : Fix a crash when deleting vtk-python objects.
  kitware/VTK@e76153 : BUG: vtkProp3D - Wrong transformation used to restore original parameters

Following discussion with Berk [1], the OpenGL backend remains the
same and the corresponding changes are backward compatible. (There are
some exceptions but Slicer code base is not impacted)

[1] http://www.kitware.com/blog/home/post/703

For a complete list of changes:

  git shortlog 169b37..0d4202 --no-merges

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23512 3bd1e089-480b-0410-8dfb-8563597acbee