ENH: Always save transforms as FromParent. See #3788
Kept readWriteTransformAsToParent flag just for backward
compatibility, when reading an old scene.

If in the future we want to save the ToParent transform or
both FromParent and ToParent then the flag has to be added
to the storage node (and not the transform node).

From: Andras Lasso <lasso@queensu.ca>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23582 3bd1e089-480b-0410-8dfb-8563597acbee