BUG: #3807 fix crashes when cloning non-scalar volumes
Since nodes are strongly typed, it's important to get the right
subclass when making the copy.  Unfortunately New is not virtual
and there's currently not a good way to make this code polymorphic.
We should investigate doing a sweeping change to MRML so that
the vtkStandardNewMacro is used appropriately but that is a bigger
change than is needed to fix this particular bug.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23590 3bd1e089-480b-0410-8dfb-8563597acbee