BUG: Ensure vtkGUISupportQtOpenGL is packaged on MacOSX. Fixes #3872
This commit add vtkGUISupportQtOpenGL to the VTK component list so that
it show up as a dependency when fixup script are used to create the MacOSX
package.

This is required because "find_package(VTK..)" is used in SlicerConfig.cmake
without specifying a list of component and the complete list of VTK libraries
is used to build extensions.

Ideally, the same list of components should also be used when calling
find_package(VTK ...) from SlicerConfig.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23755 3bd1e089-480b-0410-8dfb-8563597acbee