BUG: user loaded color files don't need to be singletons
The load data dialogue uses the color logic methods to load in
color files. These methods were originally written to load the
built in color files and have variables set that need to be
unset for user files. The singleton tag needs to be unset for
user loaded files or else they won't be cleared out of the
scene on close and can cause unexpected behaviour when
saved in MRBs.



git-svn-id: http://svn.slicer.org/Slicer4/trunk@23923 3bd1e089-480b-0410-8dfb-8563597acbee