BUG: fix restoring voxel based vs interpolation
When switching from a parameter node that had voxel based
turned on to one that didn't, the widget wasn't updating. This was
because of the way the radio buttons were set to toggle:
If a node had the voxel based flag on, the voxel based radio button had
checked set to true and the interpolation radio button had checked
set to false. Switching to the node with the voxel based flag set to false,
the previous commit was just setting the voxel based radio button to unchecked
but that was not forcing the interpolation radio button to be checked.
This change explicitly sets the checked flag to true for the mode, which
results in the other radio button in the group being set to unchecked.

Issue #3947

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23974 3bd1e089-480b-0410-8dfb-8563597acbee