BUG: Do not disable the picking in 2D, widget is never pickable #3971
This prevent the following error to appear when using 2D projection:
ERROR: In c:\d\p\slicer-440-package\vtkv6\interaction\widgets\vtkWidgetRepresentation.h, line 176
vtkPointHandleRepresentation2D (0000000010537080): Subclass should allow enable/disable picking

The fix is to remove the DisablePicking call. It's not used since for the
2D projection, the widget is never enabled (i.e. it's visible but the user
cannot interact with it). For reference, the disable call is here at
line 783 of vtkMRMLMarkupsFiducialDisplayableManager2D.cxx.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24009 3bd1e089-480b-0410-8dfb-8563597acbee