BUG: Do not fire incorrect events when closing scene
* onNodeIDChanged() was using a never set "Quantity" property.
* The unitComboBox is not exposed to the user,
so the current node id can only be changed by the selection node.
Such workflow is already handled and does not need extra signal/slot connection.

It had the side effect of resetting the number of decimals to the slider
widget of the slice views when closing the scene. That would send (queued)
signals to set the slice offset to be 0.0 AFTER (because queued connection)
the new slice offset was initialized when loading a new scene.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24013 3bd1e089-480b-0410-8dfb-8563597acbee