BUG: Fix RSNAQuantTutorial test
After downloading and extracting the dataset "dataset3_PETCT.zip"
associated with this test using url [1], it was possible to reproduce
the problem by executing these two lines in the python interpretor:

  slicer.util.loadScene('/path/to/PET_CT_pre-treatment.mrb')
  slicer.mrmlScene.Clear(0)

Considering that (1) the corresponding scene has four vtkMRMLViewNode nodes
and (2) the conventional layout is include only one vtkMRMLViewNode node,
there was a discrepancy between the method "viewCount()" and "viewWidget(int id)".

More specifically, the "viewCount()" method was returning the number of
entries in the Views ivar of the qMRMLLayoutViewFactory whereas the method
"viewWidget(int id)" was looking up the widget from within the registered
views associated with the ctkLayoutViewFactory.

It means that when "PET_CT_pre-treatment.mrb", was loaded and the scene
cleared, the method XXXX was failing to retrieve the view widget because
only one view was in the list of "registeredViews".

[1] http://slicer.kitware.com/midas3/download?items=124185

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24021 3bd1e089-480b-0410-8dfb-8563597acbee