BUG: Ensure libvtkViewsQt is packaged on MacOSX. Fixes #3973
This commit add libvtkViewsQt to the VTK component list so that it show up as a dependency when fixup script are used to create the MacOSX package.

See commit r23755 for more details on why this problem arises. This previous commit fixes the same issue for another VTK library.

From: Francois Budin <fbudin@unc.edu>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24029 3bd1e089-480b-0410-8dfb-8563597acbee