BUG: Remove obsolete NeurosurgicalPlanningTutorial QtTesting based test
QtTesting XML tests are hard to maintain.  Considering tutorial tests are
now all written in python as SelfTest, this commit removes the last QtTesting
tests.

For example, in the case of the NeurosurgicalPlanningTutorial test, the
play back is failing because the tree indices first recorded are now
invalid. Most likely because the associated widget has been updated.
Considering that (1) the XML is hard to manually update and debug and
(2) is is difficult to re-record only a section of the test starting from
the expected state, it would usually require the developer to re-record
the full test. This is time consuming and sub-optimal. With better
instrumentation or by using the QtTesting python based test (instead of
the XML), it may be worth reconsidering.

Also remove DiffusionTensorImagingTutorial related files (the associated
test has been disabled in r22388)

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24040 3bd1e089-480b-0410-8dfb-8563597acbee