STYLE: Mark qSlicerCommandOptions readonly Qt properties as CONSTANT
From Qt documentation:

The presence of the CONSTANT attribute indicates that the property value
is constant. For a given object instance, the READ method of a constant
property must return the same value every time it is called. This constant
value may be different for different instances of the object. A constant
property cannot have a WRITE method or a NOTIFY signal.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24178 3bd1e089-480b-0410-8dfb-8563597acbee