BUG: Fixed transform setting failure
When trying to set a transform that is shown in the Transforms module GUI and the new value has a component that is outside the current slider min/max range then the min/max range is updated but the transform value is not updated. The problem is that the unit module resets the value of a transform even if only the min/max range is set (it supposed to only do that if the precision is changed).

Added a check to prevent resetting of the transform value if precision is not changed.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24245 3bd1e089-480b-0410-8dfb-8563597acbee