BUG: update file list members when saving MRBs
The file lists were not getting updated when volumes with
multiple files listed were saved in the default format of nrrd.
This change saves and restores the file list around MRB saving.
Fixed the unique file name to not just append .nrrd to the end
of the current file name.
TODO: deal with the file lists in the storage nodes in the
master scene view (problem when save a scene as mrml then
save it to mrb).

Issue #3956

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24373 3bd1e089-480b-0410-8dfb-8563597acbee