BUG: avoid overwriting markup information on scene view restore
When a scene view is created in a running Slicer, the fiducial list
nodes in the scene views have correct per fiducial information
(visibility, position, etc).
When the scene is saved to disk, only one fiducial file is saved,
the current one for the main scene. The scene view nodes do
not hold per fiducial information now that all of that is saved
to disk.
The fiducials are now the same as volumes and models in terms of not
supporting data that's saved to disk changing between scene views. For
individual fiducials, this means:
position, orientation, visibility, selected, locked, label, description,
and the associated node ID

When loading a scene from disk and restoring a scene view, previously
the markups node in the Copy function would clear out the data that it
had been populated with via the storage node read. This change checks if
the scene view is restoring and the node in the scene view, from which
the data is going to get copied, has no fiducials in it, and in that
case, don't clear out the list.

Issue #3956

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24376 3bd1e089-480b-0410-8dfb-8563597acbee