STYLE: Remove version string from loadable module xml description
This will facilitate release process. It will be one less element to
update when releasing a new version of Slicer.

Rational: Some loadable module include a xml file describing their
parameter. Similarly to CLI module, this file is used on the wiki to
"automatically" document [1] the option provided by a loadable module.

The MediaWiki template responsible for transforming the XML into wiki code
is available in [2]. Since it doesn't make use of the version information,
this commit simplify the maintenance by removing it.

[1] http://www.slicer.org/slicerWiki/index.php/Documentation/Nightly/Developers/Tutorials/DocumentModule#Documentation_for_a_loadable_module

[2] http://www.slicer.org/slicerWiki/index.php?title=Template:Documentation/Nightly/module-description

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24427 3bd1e089-480b-0410-8dfb-8563597acbee