ENH: DICOM reference name made checkable
Previously the DICOM references dialog offered chacking the found but unselected references in a checkbox with no text, and added a label next to it, so the user could check/uncheck the checkbox clicking the small checkbox square, instead of reacting also to clicking on the text. By removing the label and adding text to the checkbox this inconvenience has been eliminated.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24469 3bd1e089-480b-0410-8dfb-8563597acbee