BUG: Fixed volume rendering not visible when non-volume-rendered volume is frequently changing in the scene
The issue caused volume rendering result not visible when a live 2D ultrasound image was kept updating in the scene through OpenIGTLink.

Fix: VR displayable manager is notified about the change of all volume nodes that have a VR display node, but we now only request render if the currently volume-rendered volume node is changed.


git-svn-id: http://svn.slicer.org/Slicer4/trunk@24487 3bd1e089-480b-0410-8dfb-8563597acbee