BUG: Make sure display nodes are observed before adding model to scene
This caused a subtle bug reported by Beatriz Paniagua.
Sequence to reproduce before this change:
* start fresh slicer
* make green slice visible (it appears)
* close scene
* make green slice visible (BUG: it does not appear)
* perform any other operation that changes the 3D view, green slice appears

Note that the issue only happened with the green slice model,
which importantly is the last one on the list.

Issue was because all slice models where observing their
display nodes *after* being added to the scene, meaning that
the displayable manager saw no display node on the model nodes
when it got the node added event for the model node and therefore
did not observe the node.  Other slice model nodes did not exhibit
this issue because they were being refreshed when other nodes
were added to the scene.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24506 3bd1e089-480b-0410-8dfb-8563597acbee