ENH: allow self tests to set a default message display
The default message delay for self.delayDisplay for a scripted
module test is still 1 second, but can be set to a smaller value
to speed up testing.  It's still possible to override the delay
time on a per-call basis, for example to highlight transitions
between test sections by giving a longer timeout.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24514 3bd1e089-480b-0410-8dfb-8563597acbee