BUG: Fix unitialized variable that causes py_LabelStatistics test failure
In a release build running the py_LabelStatistics test intermittently results in
a hang or memory corruption.

This commit fixes the problem by initializing a member variable in
vtkSlicerVolumesLogic. The member variable, and therefore the regression, was
introduced in r23934 ("BUG: avoid false positives on geometry check"). The
initial value matches what's described in the member variable's comment.

The unit test is updated to check the initial value of the member variable.

From: Max Smolens <max.smolens@kitware.com>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24568 3bd1e089-480b-0410-8dfb-8563597acbee