BUG: Fix crash on saving node named with a single number
A previous fix[1] to avoid crashes on windows when saving
volumes named "1:xxxx" leads to a crash when a node name
is just a single number (the first part of the check passes, but
crashes when trying to get the second character in the name string).
This fix adds a string length check so that valid node names like
"1" or "2" will not cause a crash on save.

[1] http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=24294

Issue #3991

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24722 3bd1e089-480b-0410-8dfb-8563597acbee