COMP: Fix ExternalProjectAddSource to work with non-clean build directory.
This commit ensure option associated with remote module are always
up-to-date even when changed in an existing build directory.

This particular problem was found after changing the minor version
to "5" and observing that BRAINSTools module were still building
in the lib/Slicer-4.4/cli-modules directory.

This was explained because variables like
BRAINSTools_CLI_LIBRARY_OUTPUT_DIRECTORY were not forced into the
cache, preventing the "mark_as_superbuild" function from grabbing
the up-to-date value.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24731 3bd1e089-480b-0410-8dfb-8563597acbee