ENH: Moved up vtkMRMLStorableNode in the MRML node hierarchy.
vtkMRMLStorableNode is not a children of vtkMRMLTransformable node anymore,
but directly a children of vtkMRMLNode.

This allows making a node storable without requiring it to be also
transformable. It is important for several node types (color maps, tables,
etc), which require separate storage node but are not transformable.

From: Andras Lasso <lasso@queensu.ca>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24891 3bd1e089-480b-0410-8dfb-8563597acbee