ENH: Allow adding module directories by drag-and-drop
In Settings dialog / Modules panel, it is now possible to drag a directory to the directory list to add the directory (no need to click 'Add', select the directory, and click OK).

Implemented by adding this feature to qSlicerDirectoryListView widget. Also added a designer plugin for this widget to make it show up in Qt Designer.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24914 3bd1e089-480b-0410-8dfb-8563597acbee